Revert "moved types from 'constants' to 'types' directory"
This reverts commit 458a82a5bf259758b21b201ac02d8809a6fa3cd8.
