+++ src/components/molecules/sort-selector/sort-selector.tsx
@@ -1,6 +1,5 @@
 import { SortSelectorItem } from '../../atoms';
-import { SORT_TYPES } from '../../../constants';
-import { SortTypesType } from '../../../types';
+import { SORT_TYPES, SortTypesType } from '../../../constants';
 import { useBooleanHelpers } from '../../../hooks';
 
 type SortSelectorProps = {
