+++ src/constants/cities.ts
@@ -1,3 +1,5 @@
+type CitiesType = typeof CITIES[keyof typeof CITIES];
+
 const CITIES = {
   Paris: 'Paris',
   Cologne: 'Cologne',
@@ -7,6 +9,5 @@ const CITIES = {
   Dusseldorf: 'Dusseldorf',
 } as const;
 
-export {
-  CITIES,
-};
+export default CITIES;
+export type { CitiesType };
