--- src/constants/request-status.ts
@@ -1,11 +1,0 @@
-const REQUEST_STATUS = {
-  Idle: 'Idle',
-  Loading: 'Loading',
-  Status: 'Status',
-  Error: 'Error',
-} as const;
-
-export {
-  REQUEST_STATUS,
-};
-
