+++ src/constants/sort.ts
@@ -1,3 +1,5 @@
+type SortTypesType = typeof SORT_TYPES[keyof typeof SORT_TYPES];
+
 const SORT_TYPES = {
   Popular: 'Popular',
   PriceASC: 'Price: low to high',
@@ -5,6 +7,7 @@ const SORT_TYPES = {
   Rating: 'Top rated first',
 } as const;
 
-export {
-  SORT_TYPES,
+export default SORT_TYPES;
+export type {
+  SortTypesType,
 };
