--- src/types/cities.ts
@@ -1,7 +1,0 @@
-import { CITIES } from '../constants';
-
-type CitiesType = typeof CITIES[keyof typeof CITIES];
-
-export type {
-  CitiesType,
-};
