+++ src/types/index.ts
@@ -18,15 +18,6 @@ import {
   AuthStatusType,
   AuthenticationForm,
 } from './auth.ts';
-import {
-  RequestStatusType,
-} from './request-status.ts';
-import {
-  CitiesType,
-} from './cities.ts';
-import {
-  SortTypesType,
-} from './sort.ts';
 
 export type {
   IconsAndLabelsStyleClassType,
@@ -42,7 +33,4 @@ export type {
   AuthenticatedUserType,
   AuthStatusType,
   AuthenticationForm,
-  RequestStatusType,
-  CitiesType,
-  SortTypesType,
 };
