+++ src/utils/city.ts
@@ -1,5 +1,4 @@
-import { CITIES } from '../constants';
-import { CitiesType } from '../types';
+import { CITIES, CitiesType } from '../constants';
 
 const isCity = (value: string): value is CitiesType => (
   Object.values<string>(CITIES).includes(value)
