+++ src/utils/routing.ts
@@ -1,5 +1,4 @@
-import { APP_ROUTE } from '../constants';
-import { CitiesType } from '../types';
+import { APP_ROUTE, CitiesType } from '../constants';
 
 const createMainRouteWithCity = (city: CitiesType) => APP_ROUTE.MainWithCity.replace(':city', city.toLowerCase());
